+++ src/geom/objects/triangle.rs
@@ -18,6 +18,7 @@ impl Triangle {
         note = "Use another collision library like `vek` instead; please comment on issue #552 for use-cases other libraries don't solve"
     )]
     ///Create a triangle from `Vector`s of all three points
+    #[inline]
     pub fn new(a: Vector, b: Vector, c: Vector) -> Triangle {
         Triangle { a, b, c }
     }